Moved attribute helpers to utils
